+++ src/confcom/azext_confcom/lib/templates.py
